Update tests and progress bar
